@@ -1,4 +1,4 @@
-import 'package:camelus/routes/nostr/onboarding/onboarding.dart';
+import 'package:camelus/presentation_layer/routes/nostr/onboarding/onboarding.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:flutter/material.dart';
