@@ -1,6 +1,8 @@
 import 'package:bip340/bip340.dart' as bip340;
 import 'package:camelus/helpers/helpers.dart';
 
+import '../domain_layer/entities/key_pair.dart';
+
 class Bip340 {
   final _helpers = Helpers();
 
@@ -17,6 +19,7 @@ class Bip340 {
   /// [publicKey] is a 32-bytes hex-encoded string
   /// true if the signature is valid otherwise false
   bool verify(String message, String signature, String? publicKey) {
+    if (publicKey == null) return false;
     return bip340.verify(publicKey, message, signature);
   }
 
@@ -34,36 +37,10 @@ class Bip340 {
     final privKeyHr = _helpers.encodeBech32(privKey, 'nsec');
     final pubKeyHr = _helpers.encodeBech32(pubKey, 'npub');
 
-    return KeyPair(privKey, pubKey, privKeyHr, pubKeyHr);
+    return KeyPair(
+        privateKey: privKey,
+        publicKey: pubKey,
+        privateKeyHr: privKeyHr,
+        publicKeyHr: pubKeyHr);
   }
-}
-
-class KeyPair {
-  /// [privateKey] is a 32-bytes hex-encoded string
-  final String privateKey;
-
-  /// [publicKey] is a 32-bytes hex-encoded string
-  final String publicKey;
-
-  /// [privateKeyHr] is a human readable private key e.g. nsec
-  final String privateKeyHr;
-
-  /// [publicKeyHr] is a human readable public key e.g. npub
-  final String publicKeyHr;
-
-  KeyPair(this.privateKey, this.publicKey, this.privateKeyHr, this.publicKeyHr);
-
-  Map<String, dynamic> toJson() => {
-        'privateKey': privateKey,
-        'publicKey': publicKey,
-        'privateKeyHr': privateKeyHr,
-        'publicKeyHr': publicKeyHr,
-      };
-
-  factory KeyPair.fromJson(Map<String, dynamic> json) => KeyPair(
-        json['privateKey'],
-        json['publicKey'],
-        json['privateKeyHr'],
-        json['publicKeyHr'],
-      );
 }
